refactor: move nodes to theirs own package
