+++ src/main/java/org/ua/wozzya/algo/BlockingRoughSyncQueue.java
@@ -1,5 +1,7 @@
 package org.ua.wozzya.algo;
 
+import org.ua.wozzya.algo.node.VolatileNode;
+
 import java.util.Optional;
 import java.util.Random;
 import java.util.concurrent.atomic.AtomicReference;
