+++ src/main/java/org/ua/wozzya/algo/FineGrainedList.java
@@ -1,9 +1,7 @@
 package org.ua.wozzya.algo;
 
-import scala.Int;
-import scala.concurrent.impl.FutureConvertersImpl;
+import org.ua.wozzya.algo.node.VolatileLockableNode;
 
-import java.util.Comparator;
 import java.util.Random;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
